+++ mobile/lib/screens/auth/change_email_screen.dart
@@ -196,7 +196,7 @@ class ChangeEmailScreenState extends State<ChangeEmailScreen> {
             onTap: () async {
               await _verifySentCode();
             },
-            child: nextButton('Next', _nextBtnColor),
+            child: NextButton(buttonColor: _nextBtnColor),
           ),
         ),
         Visibility(
@@ -205,7 +205,7 @@ class ChangeEmailScreenState extends State<ChangeEmailScreen> {
             onTap: () async {
               await _requestVerification();
             },
-            child: nextButton('Next', _nextBtnColor),
+            child: NextButton(buttonColor: _nextBtnColor),
           ),
         ),
         const SizedBox(
