+++ mobile/lib/screens/on_boarding/profile_setup_screen.dart
@@ -102,7 +102,7 @@ class ProfileSetupScreenState extends State<ProfileSetupScreen> {
                 onTap: () async {
                   await saveName();
                 },
-                child: nextButton('Next', nextBtnColor),
+                child: NextButton(buttonColor: nextBtnColor),
               ),
               SizedBox(
                 height: _showOptions ? 16 : 12,
