+++ mobile/lib/widgets/analytics_card.dart
@@ -23,126 +23,75 @@ import 'buttons.dart';
 import 'custom_shimmer.dart';
 import 'custom_widgets.dart';
 
-Widget analyticsCardLoading() {
-  return Container(
-    padding: const EdgeInsets.only(top: 32, bottom: 22),
-    constraints: const BoxConstraints(
-      maxHeight: 251,
-      minHeight: 251,
-    ),
-    decoration: BoxDecoration(
-      color: Colors.white,
-      borderRadius: const BorderRadius.all(Radius.circular(16.0)),
-      border: Border.all(color: Colors.transparent),
-    ),
-    child: Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.only(left: 24, right: 24),
-          child: Row(
-            children: [
-              const CircularLoadingAnimation(size: 104),
-              const SizedBox(width: 16.0),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: const [
-                    SizedContainerLoadingAnimation(
-                        height: 15, width: 139, radius: 1000),
-                    SizedBox(
-                      height: 9,
-                    ),
-                    SizedContainerLoadingAnimation(
-                        height: 10, width: 115, radius: 1000),
-                    SizedBox(
-                      height: 12,
-                    ),
-                    SizedContainerLoadingAnimation(
-                        height: 24, width: 115, radius: 1000),
-                  ],
-                ),
-              )
-            ],
-          ),
-        ),
-        const SizedBox(height: 20),
-        const Padding(
-          padding: EdgeInsets.only(left: 24, right: 24),
-          child: ContainerLoadingAnimation(height: 9, radius: 1000),
-        ),
-        const Divider(color: Color(0xffC4C4C4)),
-        const SizedBox(
-          height: 16,
-        ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: const [
-            SizedContainerLoadingAnimation(
-                height: 20, width: 105, radius: 1000),
-            SizedContainerLoadingAnimation(
-                height: 20, width: 105, radius: 1000),
+class MapAnalyticsMoreInsights extends StatelessWidget {
+  final PlaceDetails placeDetails;
+  const MapAnalyticsMoreInsights({Key? key, required this.placeDetails})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 16,
+      child: ListTile(
+        contentPadding: const EdgeInsets.only(left: 20, right: 30),
+        title: Row(
+          children: [
+            SvgPicture.asset(
+              'assets/icon/chart.svg',
+              semanticsLabel: 'chart',
+              height: 16,
+              width: 16,
+            ),
+            const SizedBox(width: 8.0),
+            Text(
+              'View More Insights',
+              style: TextStyle(fontSize: 12, color: Config.appColorBlue),
+            ),
+            const Spacer(),
+            SvgPicture.asset(
+              'assets/icon/more_arrow.svg',
+              semanticsLabel: 'more',
+              height: 6.99,
+              width: 4,
+            ),
           ],
         ),
+      ),
+    );
+  }
+}
+
+class AnalyticsMoreInsights extends StatelessWidget {
+  final PlaceDetails placeDetails;
+  const AnalyticsMoreInsights({Key? key, required this.placeDetails})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        SvgPicture.asset(
+          'assets/icon/chart.svg',
+          semanticsLabel: 'chart',
+          height: 16,
+          width: 16,
+        ),
+        const SizedBox(width: 8.0),
+        Text(
+          'View More Insights',
+          style: CustomTextStyle.caption4(context)
+              ?.copyWith(color: Config.appColorBlue),
+        ),
+        const Spacer(),
+        SvgPicture.asset(
+          'assets/icon/more_arrow.svg',
+          semanticsLabel: 'more',
+          height: 6.99,
+          width: 4,
+        ),
       ],
-    ),
-  );
-}
-
-Widget mapMoreInsightsWidget(PlaceDetails placeDetails) {
-  return SizedBox(
-    height: 16,
-    child: ListTile(
-      contentPadding: const EdgeInsets.only(left: 20, right: 30),
-      title: Row(
-        children: [
-          SvgPicture.asset(
-            'assets/icon/chart.svg',
-            semanticsLabel: 'chart',
-            height: 16,
-            width: 16,
-          ),
-          const SizedBox(width: 8.0),
-          Text(
-            'View More Insights',
-            style: TextStyle(fontSize: 12, color: Config.appColorBlue),
-          ),
-          const Spacer(),
-          SvgPicture.asset(
-            'assets/icon/more_arrow.svg',
-            semanticsLabel: 'more',
-            height: 6.99,
-            width: 4,
-          ),
-        ],
-      ),
-    ),
-  );
-}
-
-Widget moreInsightsWidget(PlaceDetails placeDetails, context) {
-  return Row(
-    children: [
-      SvgPicture.asset(
-        'assets/icon/chart.svg',
-        semanticsLabel: 'chart',
-        height: 16,
-        width: 16,
-      ),
-      const SizedBox(width: 8.0),
-      Text(
-        'View More Insights',
-        style: CustomTextStyle.caption4(context)
-            ?.copyWith(color: Config.appColorBlue),
-      ),
-      const Spacer(),
-      SvgPicture.asset(
-        'assets/icon/more_arrow.svg',
-        semanticsLabel: 'more',
-        height: 6.99,
-        width: 4,
-      ),
-    ],
-  );
+    );
+  }
 }
 
 class AnalyticsCard extends StatefulWidget {
@@ -321,7 +270,8 @@ class _AnalyticsCardState extends State<AnalyticsCard> {
                     const SizedBox(height: 30),
                     Padding(
                       padding: const EdgeInsets.only(left: 24, right: 24),
-                      child: moreInsightsWidget(widget.placeDetails, context),
+                      child: AnalyticsMoreInsights(
+                          placeDetails: widget.placeDetails),
                     ),
                     const SizedBox(height: 12),
                     const Divider(
@@ -577,8 +527,9 @@ class _MapAnalyticsCardState extends State<MapAnalyticsCard> {
 
                         const SizedBox(height: 20),
                         // Analytics
-                        mapMoreInsightsWidget(
-                          PlaceDetails.measurementToPLace(widget.measurement),
+                        MapAnalyticsMoreInsights(
+                          placeDetails: PlaceDetails.measurementToPLace(
+                              widget.measurement),
                         ),
                       ],
                     ),
@@ -666,6 +617,215 @@ class _MapAnalyticsCardState extends State<MapAnalyticsCard> {
         _showHeartAnimation = false;
       });
     });
+    await _appService.updateFavouritePlace(widget.placeDetails, context);
+  }
+}
+
+class MiniAnalyticsCard extends StatefulWidget {
+  final PlaceDetails placeDetails;
+
+  const MiniAnalyticsCard(this.placeDetails, {Key? key}) : super(key: key);
+
+  @override
+  _MiniAnalyticsCard createState() => _MiniAnalyticsCard();
+}
+
+class _MiniAnalyticsCard extends State<MiniAnalyticsCard> {
+  late Measurement measurement;
+  bool showHeartAnimation = false;
+  bool isNull = true;
+
+  final AppService _appService = AppService();
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {
+        Navigator.push(context, MaterialPageRoute(builder: (context) {
+          return InsightsPage(widget.placeDetails);
+        }));
+      },
+      child: Padding(
+        padding: const EdgeInsets.fromLTRB(0.0, 8.0, 0.0, 8.0),
+        child: Container(
+            decoration: BoxDecoration(
+                color: Colors.white,
+                borderRadius: const BorderRadius.all(Radius.circular(8.0)),
+                border: Border.all(color: Colors.transparent)),
+            child: Column(
+              children: [
+                const SizedBox(
+                  height: 24,
+                ),
+                Container(
+                  padding: const EdgeInsets.only(left: 32, right: 32),
+                  child: Row(
+                    children: [
+                      if (!isNull)
+                        MiniAnalyticsAvatar(measurement: measurement),
+                      if (isNull) const CircularLoadingAnimation(size: 40),
+                      const SizedBox(
+                        width: 12,
+                      ),
+                      Expanded(
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          mainAxisAlignment: MainAxisAlignment.start,
+                          children: [
+                            AutoSizeText(
+                              widget.placeDetails.name,
+                              maxLines: 1,
+                              overflow: TextOverflow.ellipsis,
+                              style: CustomTextStyle.headline8(context),
+                            ),
+                            AutoSizeText(
+                              widget.placeDetails.location,
+                              maxLines: 1,
+                              overflow: TextOverflow.ellipsis,
+                              style: CustomTextStyle.bodyText4(context)
+                                  ?.copyWith(
+                                      color: Config.appColorBlack
+                                          .withOpacity(0.3)),
+                            ),
+                          ],
+                        ),
+                      ),
+                      const SizedBox(
+                        width: 12,
+                      ),
+                      Consumer<PlaceDetailsModel>(
+                        builder: (context, placeDetailsModel, child) {
+                          return GestureDetector(
+                              onTap: () async {
+                                updateFavPlace();
+                              },
+                              child: getHeartIcon());
+                        },
+                      )
+                    ],
+                  ),
+                ),
+                const SizedBox(
+                  height: 24,
+                ),
+                const Divider(color: Color(0xffC4C4C4)),
+                const SizedBox(
+                  height: 11,
+                ),
+                Container(
+                  padding: const EdgeInsets.only(left: 32, right: 32),
+                  child: Row(
+                    children: [
+                      Container(
+                        height: 16,
+                        width: 16,
+                        decoration: BoxDecoration(
+                            color: Config.appColorBlue,
+                            borderRadius:
+                                const BorderRadius.all(Radius.circular(3.0)),
+                            border: Border.all(color: Colors.transparent)),
+                        child: const Icon(
+                          Icons.bar_chart,
+                          size: 14,
+                          color: Colors.white,
+                        ),
+                      ),
+                      const SizedBox(width: 8.0),
+                      Text(
+                        'View More Insights',
+                        style: CustomTextStyle.caption3(context)
+                            ?.copyWith(color: Config.appColorBlue),
+                      ),
+                      const Spacer(),
+                      Container(
+                        height: 16,
+                        width: 16,
+                        padding: const EdgeInsets.all(5),
+                        decoration: BoxDecoration(
+                            color: Config.appColorPaleBlue,
+                            borderRadius:
+                                const BorderRadius.all(Radius.circular(3.0)),
+                            border: Border.all(color: Colors.transparent)),
+                        child: SvgPicture.asset(
+                          'assets/icon/more_arrow.svg',
+                          semanticsLabel: 'more',
+                          height: 6.99,
+                          width: 4,
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+                const SizedBox(
+                  height: 20,
+                ),
+              ],
+            )),
+      ),
+    );
+  }
+
+  Widget getHeartIcon() {
+    if (showHeartAnimation) {
+      return SizedBox(
+        height: 16.67,
+        width: 16.67,
+        child: Lottie.asset('assets/lottie/animated_heart.json',
+            repeat: false, reverse: false, animate: true, fit: BoxFit.cover),
+      );
+    }
+
+    return Consumer<PlaceDetailsModel>(
+      builder: (context, placeDetailsModel, child) {
+        if (PlaceDetails.isFavouritePlace(
+            placeDetailsModel.favouritePlaces, widget.placeDetails)) {
+          return SvgPicture.asset(
+            'assets/icon/heart.svg',
+            semanticsLabel: 'Favorite',
+            height: 16.67,
+            width: 16.67,
+          );
+        }
+        return SvgPicture.asset(
+          'assets/icon/heart_dislike.svg',
+          semanticsLabel: 'Favorite',
+          height: 16.67,
+          width: 16.67,
+        );
+      },
+    );
+  }
+
+  void getMeasurement() {
+    _appService.dbHelper
+        .getMeasurement(widget.placeDetails.siteId)
+        .then((value) => {
+              if (value != null && mounted)
+                {
+                  setState(() {
+                    measurement = value;
+                    isNull = false;
+                  })
+                }
+            });
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    getMeasurement();
+  }
+
+  void updateFavPlace() async {
+    setState(() {
+      showHeartAnimation = true;
+    });
+    Future.delayed(const Duration(seconds: 2), () {
+      setState(() {
+        showHeartAnimation = false;
+      });
+    });
+
     await _appService.updateFavouritePlace(widget.placeDetails, context);
   }
 }
