+++ mobile/lib/widgets/buttons.dart
@@ -3,34 +3,42 @@ import 'package:flutter_svg/svg.dart';
 
 import '../constants/config.dart';
 
-Widget nextButton(String text, Color buttonColor) {
-  return Container(
-    height: 48,
-    constraints: const BoxConstraints(minWidth: double.infinity),
-    decoration: BoxDecoration(
-        color: buttonColor,
-        borderRadius: const BorderRadius.all(Radius.circular(8.0))),
-    child: Row(
-      crossAxisAlignment: CrossAxisAlignment.center,
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Text(
-          text,
-          style: const TextStyle(
-              color: Colors.white, fontSize: 14, letterSpacing: 16 * -0.022),
-        ),
-        const SizedBox(
-          width: 11,
-        ),
-        SvgPicture.asset(
-          'assets/icon/next_arrow.svg',
-          semanticsLabel: 'Share',
-          height: 17.42,
-          width: 10.9,
-        ),
-      ],
-    ),
-  );
+class NextButton extends StatelessWidget {
+  final String? text;
+  final Color buttonColor;
+  const NextButton({Key? key, required this.buttonColor, this.text})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 48,
+      constraints: const BoxConstraints(minWidth: double.infinity),
+      decoration: BoxDecoration(
+          color: buttonColor,
+          borderRadius: const BorderRadius.all(Radius.circular(8.0))),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(
+            text ?? 'Next',
+            style: const TextStyle(
+                color: Colors.white, fontSize: 14, letterSpacing: 16 * -0.022),
+          ),
+          const SizedBox(
+            width: 11,
+          ),
+          SvgPicture.asset(
+            'assets/icon/next_arrow.svg',
+            semanticsLabel: 'Share',
+            height: 17.42,
+            width: 10.9,
+          ),
+        ],
+      ),
+    );
+  }
 }
 
 class AppBackButton extends StatelessWidget {
