+++ mobile/lib/widgets/custom_shimmer.dart
@@ -124,3 +124,73 @@ class TextLoadingAnimation extends StatelessWidget {
     );
   }
 }
+
+class AnalyticsCardLoading extends StatelessWidget {
+  const AnalyticsCardLoading({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.only(top: 32, bottom: 22),
+      constraints: const BoxConstraints(
+        maxHeight: 251,
+        minHeight: 251,
+      ),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: const BorderRadius.all(Radius.circular(16.0)),
+        border: Border.all(color: Colors.transparent),
+      ),
+      child: Column(
+        children: [
+          Padding(
+            padding: const EdgeInsets.only(left: 24, right: 24),
+            child: Row(
+              children: [
+                const CircularLoadingAnimation(size: 104),
+                const SizedBox(width: 16.0),
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: const [
+                      SizedContainerLoadingAnimation(
+                          height: 15, width: 139, radius: 1000),
+                      SizedBox(
+                        height: 9,
+                      ),
+                      SizedContainerLoadingAnimation(
+                          height: 10, width: 115, radius: 1000),
+                      SizedBox(
+                        height: 12,
+                      ),
+                      SizedContainerLoadingAnimation(
+                          height: 24, width: 115, radius: 1000),
+                    ],
+                  ),
+                )
+              ],
+            ),
+          ),
+          const SizedBox(height: 20),
+          const Padding(
+            padding: EdgeInsets.only(left: 24, right: 24),
+            child: ContainerLoadingAnimation(height: 9, radius: 1000),
+          ),
+          const Divider(color: Color(0xffC4C4C4)),
+          const SizedBox(
+            height: 16,
+          ),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            children: const [
+              SizedContainerLoadingAnimation(
+                  height: 20, width: 105, radius: 1000),
+              SizedContainerLoadingAnimation(
+                  height: 20, width: 105, radius: 1000),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
